fixed dependencies
