+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/common/config/GeneralConfig.kt
@@ -1,26 +1,45 @@
 package ac.uk.ebi.biostd.common.config
 
 import ac.uk.ebi.biostd.common.properties.ApplicationProperties
+import ac.uk.ebi.biostd.files.FileConfig
 import ac.uk.ebi.biostd.persistence.common.service.SubmissionFilesPersistenceService
+import ac.uk.ebi.biostd.persistence.doc.integration.MongoDbServicesConfig
 import ac.uk.ebi.biostd.stats.web.TempFileGenerator
 import ac.uk.ebi.biostd.submission.service.FileSourcesService
 import ebi.ac.uk.ftp.FtpClient
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Import
 import uk.ac.ebi.fire.client.integration.web.FireClient
+import uk.ac.ebi.fire.client.integration.web.FireClientFactory
+import uk.ac.ebi.fire.client.integration.web.FireConfig
+import uk.ac.ebi.fire.client.integration.web.RetryConfig
+import uk.ac.ebi.fire.client.integration.web.S3Config
 import uk.ac.ebi.io.builder.FilesSourceListBuilder
 import uk.ac.ebi.io.config.FilesSourceConfig
 import uk.ac.ebi.serialization.common.FilesResolver
 import java.io.File
 import java.nio.file.Paths
+import kotlin.time.Duration.Companion.minutes
 
 @Configuration
+@Import(MongoDbServicesConfig::class)
 @EnableConfigurationProperties(ApplicationProperties::class)
 internal class GeneralConfig {
     @Bean
     fun tempFileGenerator(properties: ApplicationProperties): TempFileGenerator =
         TempFileGenerator(properties.tempDirPath)
+
+    @Bean
+    fun filesSourceListBuilder(config: FilesSourceConfig): FilesSourceListBuilder = config.filesSourceListBuilder()
+
+    @Bean
+    fun fileSourcesService(builder: FilesSourceListBuilder): FileSourcesService = FileSourcesService(builder)
+
+    @Bean
+    fun extFilesResolver(properties: ApplicationProperties): FilesResolver =
+        FilesResolver(File(properties.requestFilesPath))
 
     @Bean
     fun filesSourceConfig(
@@ -32,12 +51,32 @@ internal class GeneralConfig {
         FilesSourceConfig(Paths.get(applicationProperties.submissionPath), fireClient, filesRepo, ftpClient)
 
     @Bean
-    fun filesSourceListBuilder(config: FilesSourceConfig): FilesSourceListBuilder = config.filesSourceListBuilder()
+    fun fireClient(properties: ApplicationProperties): FireClient {
+        val fireProps = properties.fire
+        val retryProps = fireProps.retry
+        return FireClientFactory.create(
+            FireConfig(
+                fireHost = fireProps.host,
+                fireVersion = fireProps.version,
+                username = fireProps.username,
+                password = properties.fire.password
+            ),
+            S3Config(
+                accessKey = fireProps.s3.accessKey,
+                secretKey = fireProps.s3.secretKey,
+                region = fireProps.s3.region,
+                endpoint = fireProps.s3.endpoint,
+                bucket = fireProps.s3.bucket
+            ),
+            RetryConfig(
+                maxAttempts = retryProps.maxAttempts,
+                initialInterval = retryProps.initialInterval,
+                multiplier = retryProps.multiplier,
+                maxInterval = retryProps.maxInterval.minutes.inWholeMilliseconds,
+            )
+        )
+    }
 
     @Bean
-    fun fileSourcesService(builder: FilesSourceListBuilder): FileSourcesService = FileSourcesService(builder)
-
-    @Bean
-    fun extFilesResolver(properties: ApplicationProperties): FilesResolver =
-        FilesResolver(File(properties.requestFilesPath))
+    fun ftpClient(properties: ApplicationProperties) = FileConfig.ftpClient(properties.security.filesProperties)
 }
