+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/files/FileConfig.kt
@@ -1,6 +1,5 @@
 package ac.uk.ebi.biostd.files
 
-import ac.uk.ebi.biostd.common.properties.ApplicationProperties
 import ac.uk.ebi.biostd.common.properties.FilesProperties
 import ac.uk.ebi.biostd.files.service.FileServiceFactory
 import ac.uk.ebi.biostd.files.web.common.FilesMapper
@@ -13,9 +12,6 @@ class FileConfig {
 
     @Bean
     fun fileMapper() = FilesMapper()
-
-    @Bean
-    fun ftpClient(properties: ApplicationProperties) = ftpClient(properties.security.filesProperties)
 
     @Bean
     fun fileServiceFactory(ftpClient: FtpClient) = FileServiceFactory(ftpClient)
